Convert camera variables to doubles

+++ src/main/java/frc/robot/Limelight.java
@@ -8,8 +8,8 @@ import edu.wpi.first.networktables.NetworkTableInstance;
 public class Limelight {
 
     // TODO: Get proper dimension varibles.
-    private final int IMG_WIDTH = 320;
-    private final int CAM_FOV = 59.6;
+    private final double IMG_WIDTH = 320;
+    private final double CAM_FOV = 59.6;
 
     private double tv;
     private double tx;
